Rename CLI and update documentation for language-agnostic support
- Rename loquilex/cli/live_en_to_zh.py -> loquilex/cli/live.py
- Create backward compat shim at old location with deprecation warning
- Update pyproject.toml entry point: loquilex-live -> loquilex.cli.live:main
- Update supervisor.py subprocess call to use new module path
- Update tests to import from new location
- Add comprehensive MIGRATION.md guide with examples
- Update README with language pair documentation and CLI usage examples
- Update CLI docstring to reflect multi-language support

Backward compatibility:
- Old import path still works: from loquilex.cli.live_en_to_zh import main
- Emits DeprecationWarning directing users to new path
- Shim will be removed in version 1.0.0

All 411 tests pass. Lint and typecheck clean.
